Refactor classes to new project
- update build.sbt
- refactor java/scala/test classes
