+++ utils/src/main/java/io/delta/utils/ParquetSchemaConverter.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package io.delta.standalone.util;
+package io.delta.utils;
 
+import io.delta.utils.internal.SparkToParquetSchemaConverter;
 import org.apache.parquet.schema.MessageType;
 
 import io.delta.standalone.types.StructType;
-import io.delta.standalone.internal.util.SparkToParquetSchemaConverter;
 
 /**
  * :: DeveloperApi ::
