+++ utils/src/test/scala/io/delta/utils/ParquetSchemaConverterSuite.scala
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package io.delta.standalone.internal
+package io.delta.utils
 
+import io.delta.utils.ParquetSchemaConverter.ParquetOutputTimestampType
 import org.apache.parquet.schema.MessageTypeParser
 import org.scalatest.FunSuite
 
 import io.delta.standalone.types._
-import io.delta.standalone.util.ParquetSchemaConverter
 
 class ParquetSchemaConverterSuite extends FunSuite {
 
@@ -29,8 +29,7 @@ class ParquetSchemaConverterSuite extends FunSuite {
       sqlSchema: StructType,
       parquetSchema: String,
       writeLegacyParquetFormat: Boolean,
-      outputTimestampType: ParquetSchemaConverter.ParquetOutputTimestampType =
-      ParquetSchemaConverter.ParquetOutputTimestampType.INT96): Unit = {
+      outputTimestampType: ParquetOutputTimestampType = ParquetOutputTimestampType.INT96): Unit = {
 
     test(s"sql => parquet: $testName") {
       val actual = ParquetSchemaConverter.deltaToParquet(
